General update v.1.0.0
"In this first version, the codes are uploaded to train the model using the script called 'model.py', where the best model is obtained after searching for the best parameters of various tree-based algorithms.

The data stored in the 'data/Fish.csv' folder is used as input for the 'model.py' script, and we will obtain a file called 'best_model.json', which will contain the information of the model with the best hyperparameters obtained for this case study.

Finally, the 'prediction.py' script contains the information to apply the model. It takes 'best_model.json' as an input parameter and creates a file with the encoded classes called 'classes.npy'. Within this same script, a small example of usage is provided, including data input, adjustment, and application of the model."